+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/angular
 
+## 3.0.40
+
+### Patch Changes
+
+- Updated dependencies [c3782a6]
+  - ai@7.0.40
+
+## 3.0.39
+
+### Patch Changes
+
+- Updated dependencies [0c464d9]
+- Updated dependencies [09a52cb]
+- Updated dependencies [c49380c]
+  - @ai-sdk/provider-utils@5.0.14
+  - ai@7.0.39
+
 ## 3.0.38
 
 ### Patch Changes