+++ packages/fireworks/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/fireworks
 
+## 3.0.17
+
+### Patch Changes
+
+- Updated dependencies [0c464d9]
+- Updated dependencies [c49380c]
+  - @ai-sdk/provider-utils@5.0.14
+  - @ai-sdk/openai-compatible@3.0.16
+
 ## 3.0.16
 
 ### Patch Changes